Intial drop of the Vulkan compute pipeline
This commit adds classes for various Vulkan objects and brings up
a minimal benchmark which just performs buffer copy. Various
components are wired up, including compiling GLSL using shader
tools, using SPIRV-Reflect to refect on the generated SPIR-V to
construct pipeline layouts and others. Adding a benchmark binary
is expected to be just as simple as throwing in a shader, and then
calling a few APIs that abstracts away most boilerplate coming
from vanilla Vulkan. The modelling on the command buffer side is
kept with relative low abstraction given that is the most
interesting part we'd like to interact with, except for shaders.

The benchmarking side of the infrastrucutre is yet to be improved
to provide better utilities and support.