Add support for generating shader permutation
Normally we'd like to use specialization constant to define shader
parameters that can be finalized at driver compilation time to
allow customizing shaders. But specialization constant is not
always properly supported everywhere; for example, driver might not
be able to properly loop unroll when they appear as the loop bounds.
For such cases, we still need to go back to the old-fashioned way
of generating shader permutations.

This commit adds a uvkc_glsl_shader_permutation CMake function.
Compared to uvkc_glsl_shader_instance, it takes a PERMUTATION
parameter, which one can specify a bunch of macros and their
choices and each production of those choices will be used to invoke
the shader compiler once. Then all compiled SPIR-V blobs are
collected into the same file so one can include in the Vulkan
benchmark application.