Add library for handling vulkan dynamic symbols
In Android 10, Vulkan applications running under /data/local/tmp
with adb might not be able to discover the Vulkan ICD, due to an
Android linker namespace issue. So, other than packing the
benchmark as a full-blown Android app, which requires quite
non-trivial efforts, a known workaround is to directly dynamically
link to the Vulkan ICD, by renaming it as libvulkan.so, to bypass
the Android Vulkan loader. That means, dynamic loading. Thus, this
commit, which is basically taken from IREE.