Add a tiled conv2d benchmark (#3)
This commit adds a tiled implementation of 2-D convolution. It tiles along
both the output width dimension and output channel dimension. Each tile
is assigned to one workgroup, which is further distributed among the
invocations in the workgroup.

Invocations perform 4-element vector load of the filter in a cyclic way.
Each invocation calculates the full result of the output element it covers
so no synchronization. Invocations write out the output elements in a cyclic
way.