Add benchmark sample for vector times matrix transposed
This adds benchmarks for `vmt`, with very similar supporting structure
to the existing `mmt` benchmark, but with different strategies tuned for
matvec. This add three strategies:

1) Treat it like a reduction with one workgroup per row, relying on
   cache to get reuse of the vector.
2) Copy the vector to shared memory using all threads in the workgroup
   and then process N0 rows per workgroup, with WG_Y | N0 threadgroups.
3) Use a fixed number of workgroups and each workgroup strides the
   full problem space. This should limit the overhead of setting up the
   vector in shared memory, as well as improves scheduling overhead.

Currently, the best configuration for each of the above three strategies
are in the same performance ballpark (~20us for a 4096 * 4096x4096 matvec
on an AMD 7900xtx).