Add RenderDoc integration
Use the RenderDoc API programatically start a capture. We need this
because uVkCompute does not present any frames and frame-oriented tools
do not know when to start on their own.

The initial implementation supports linux only. The RenderDoc header is
vendored under `/third_party`.

RenderDoc capture is enabled with a new `--enable_renderdoc` flag. This
is coarse-grained: we start a signle capture for all of the benchmarks,
so you may want to combine it with a `--benchmark_filter` to profile
only some configurations of interest.
In the future we may want to make it more fine-grained, e.g., per
registered benchmark, but I was not able to find a clean way to hook
into the loop over the final benchmark list.

Note that this transitively enables RGP (Radeon GPU Profiler) support,
although without instruction timing.