Fix memory benchmarks for unexpected gl_SubgroupSize
Some Intel GPUs have flexible subgroup sizes.
subgroupSize can be 32 but minSubgroupSize can be smaller.
In this case, unless you forcibly control the subgroup size
at pipeline creation time, gl_SubgroupSize will report 32 but
the actual number of invocations in the subgroup may be 8.

In the memory benchmarks, use a bitcount of the ballot to compute
the dynamic (actual) size of the subgroup.  The alternative is
to use the much more recent (and less portable) subgroup size
control extension.

Fixes: #43