Fix subgroup_arithmetic benchmark for flexible subgroup sizes
Set up source data values according to the subgroupSize device property.
But inside the shader, measure the actual subgroup size, and pass it
out of the shader. Then, adjust the verification logic to take that
actual subgroup size into account.

This still does the same amount of arithmetic as in the original
code, and it does it in the same shape.

Fixes: #45